+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added `contour` function
 ### Fixed
 - Holes in boolean results could lead to incorrect geometry in specific cases
+- Bug in boolean operations resulting in self-intersecting polygons
 
 ## 0.6.1 - 2021-07-03
 ### Fixed